@@ -20,6 +20,7 @@ import 'package:appflowy_backend/protobuf/flowy-folder/view.pb.dart';
 import 'package:appflowy_board/appflowy_board.dart';
 import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:easy_localization/easy_localization.dart';
+import 'package:flowy_infra/theme_extension.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flowy_infra_ui/style_widget/hover.dart';
 import 'package:flowy_infra_ui/widget/error_page.dart';
@@ -671,7 +672,7 @@ class _BoardCardState extends State<_BoardCard> {
                     ? const Color(0x0F1F2329)
                     : const Color(0x0FEFF4FB),
                 foregroundColorOnHover:
-                    Theme.of(context).colorScheme.onBackground,
+                    AFThemeExtension.of(context).onBackground,
               ),
             ),
             onStartEditing: () =>
