@@ -8,6 +8,7 @@ import 'package:appflowy/shared/feedback_gesture_detector.dart';
 import 'package:appflowy/util/font_family_extension.dart';
 import 'package:appflowy/workspace/application/settings/appearance/base_appearance.dart';
 import 'package:easy_localization/easy_localization.dart';
+import 'package:flowy_infra/theme_extension.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -208,8 +209,7 @@ class _FontButton extends StatelessWidget {
       showDoneButton: true,
       showHeader: true,
       title: LocaleKeys.titleBar_font.tr(),
-      backgroundColor: Theme.of(context).colorScheme.background,
-      isScrollControlled: true,
+      backgroundColor: AFThemeExtension.of(context).background,
       enableDraggableScrollable: true,
       minChildSize: 0.6,
       initialChildSize: 0.61,
