@@ -1,5 +1,3 @@
-import 'package:flutter/material.dart';
-
 import 'package:appflowy/env/cloud_env.dart';
 import 'package:appflowy/env/env.dart';
 import 'package:appflowy/generated/flowy_svgs.g.dart';
@@ -15,7 +13,9 @@ import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:appflowy_popover/appflowy_popover.dart';
 import 'package:collection/collection.dart';
 import 'package:easy_localization/easy_localization.dart';
+import 'package:flowy_infra/theme_extension.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
 
@@ -127,7 +127,7 @@ class CloudTypeSwitcher extends StatelessWidget {
             child: FlowyTextButton(
               padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 6),
               titleFromCloudType(cloudType),
-              fontColor: Theme.of(context).colorScheme.onBackground,
+              fontColor: AFThemeExtension.of(context).onBackground,
               fillColor: Colors.transparent,
               onPressed: () {},
             ),
@@ -159,7 +159,6 @@ class CloudTypeSwitcher extends StatelessWidget {
                 showHeader: true,
                 showDragHandle: true,
                 showDivider: false,
-                showCloseButton: false,
                 title: LocaleKeys.settings_menu_cloudServerType.tr(),
                 builder: (context) {
                   return Column(
