@@ -1,5 +1,6 @@
 import 'package:appflowy/generated/flowy_svgs.g.dart';
 import 'package:appflowy/workspace/presentation/settings/widgets/theme_upload/theme_upload.dart';
+import 'package:flowy_infra/theme_extension.dart';
 import 'package:flowy_infra_ui/style_widget/text.dart';
 import 'package:flutter/material.dart';
 
@@ -24,7 +25,7 @@ class ThemeUploadFailureWidget extends StatelessWidget {
           FlowySvg(
             FlowySvgs.close_m,
             size: ThemeUploadWidget.iconSize,
-            color: Theme.of(context).colorScheme.onBackground,
+            color: AFThemeExtension.of(context).onBackground,
           ),
           FlowyText.medium(
             errorMessage,
