@@ -8,6 +8,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flowy_infra_ui/widget/buttons/secondary_button.dart';
 import 'package:flowy_infra_ui/widget/error_page.dart';
+import 'package:flowy_infra/theme_extension.dart';
 
 class ThemeUploadLearnMoreButton extends StatelessWidget {
   const ThemeUploadLearnMoreButton({super.key});
@@ -21,7 +22,7 @@ class ThemeUploadLearnMoreButton extends StatelessWidget {
       height: ThemeUploadWidget.buttonSize.height,
       child: IntrinsicWidth(
         child: SecondaryButton(
-          outlineColor: Theme.of(context).colorScheme.onBackground,
+          outlineColor: AFThemeExtension.of(context).onBackground,
           child: Padding(
             padding: const EdgeInsets.symmetric(horizontal: 8),
             child: FlowyText.medium(
