@@ -2,6 +2,7 @@ import 'package:appflowy/generated/flowy_svgs.g.dart';
 import 'package:appflowy/generated/locale_keys.g.dart';
 import 'package:appflowy/workspace/presentation/settings/widgets/theme_upload/theme_upload.dart';
 import 'package:easy_localization/easy_localization.dart';
+import 'package:flowy_infra/theme_extension.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flutter/material.dart';
 
@@ -13,7 +14,7 @@ class UploadNewThemeWidget extends StatelessWidget {
     return Container(
       color: Theme.of(context)
           .colorScheme
-          .background
+          .surface
           .withOpacity(ThemeUploadWidget.fadeOpacity),
       padding: ThemeUploadWidget.padding,
       child: Column(
@@ -23,7 +24,7 @@ class UploadNewThemeWidget extends StatelessWidget {
           FlowySvg(
             FlowySvgs.folder_m,
             size: ThemeUploadWidget.iconSize,
-            color: Theme.of(context).colorScheme.onBackground,
+            color: AFThemeExtension.of(context).onBackground,
           ),
           FlowyText.medium(
             LocaleKeys.settings_appearance_themeUpload_description.tr(),
