@@ -1,22 +1,19 @@
 // Basis
+export '/widget/separated_flex.dart';
+export '/widget/spacing.dart';
 export 'basis.dart';
-
-// Keyboard
-export 'src/keyboard/keyboard_visibility_detector.dart';
-
+export 'src/flowy_overlay/appflowy_popover.dart';
+export 'src/flowy_overlay/flowy_dialog.dart';
 // Overlay
 export 'src/flowy_overlay/flowy_overlay.dart';
 export 'src/flowy_overlay/list_overlay.dart';
 export 'src/flowy_overlay/option_overlay.dart';
-export 'src/flowy_overlay/flowy_dialog.dart';
-export 'src/flowy_overlay/appflowy_popover.dart';
+// Keyboard
+export 'src/keyboard/keyboard_visibility_detector.dart';
+export 'style_widget/button.dart';
+export 'style_widget/color_picker.dart';
+export 'style_widget/icon_button.dart';
+export 'style_widget/scrolling/styled_list.dart';
+export 'style_widget/scrolling/styled_scroll_bar.dart';
 export 'style_widget/text.dart';
 export 'style_widget/text_field.dart';
-
-export 'style_widget/button.dart';
-export 'style_widget/icon_button.dart';
-export 'style_widget/scrolling/styled_scroll_bar.dart';
-export '/widget/spacing.dart';
-export '/widget/separated_flex.dart';
-export 'style_widget/scrolling/styled_list.dart';
-export 'style_widget/color_picker.dart';
